platform--LDAP/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
python-ldap==3.4.3
Flask-Cors==3.0.10
requests==2.28.2
six==1.16.0
